+++ ajax/generationfields.php
@@ -1,0 +1,46 @@
+<?php
+
+/**
+ * -------------------------------------------------------------------------
+ * Order plugin for GLPI
+ * -------------------------------------------------------------------------
+ *
+ * LICENSE
+ *
+ * This file is part of Order.
+ *
+ * Order is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Order is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Order. If not, see <http://www.gnu.org/licenses/>.
+ * -------------------------------------------------------------------------
+ * @copyright Copyright (C) 2009-2023 by Order plugin team.
+ * @license   GPLv3 https://www.gnu.org/licenses/gpl-3.0.html
+ * @link      https://github.com/pluginsGLPI/order
+ * -------------------------------------------------------------------------
+ */
+
+/**
+ * Options of the mappable-field dropdown, for one itemtype.
+ */
+
+Session::checkLoginUser();
+Session::checkRight('config', UPDATE);
+
+$itemtype = (string) ($_GET['itemtype'] ?? '');
+$fields   = PluginOrderGenerationField::getAvailableFields($itemtype);
+
+echo "<option value=''>-----</option>";
+foreach ($fields as $field => $label) {
+    echo "<option value='" . htmlescape($field) . "'>"
+       . htmlescape(sprintf('%s (%s)', $label, $field))
+       . "</option>";
+}