+++ front/generationfield.form.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * -------------------------------------------------------------------------
+ * Order plugin for GLPI
+ * -------------------------------------------------------------------------
+ *
+ * LICENSE
+ *
+ * This file is part of Order.
+ *
+ * Order is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Order is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Order. If not, see <http://www.gnu.org/licenses/>.
+ * -------------------------------------------------------------------------
+ * @copyright Copyright (C) 2009-2023 by Order plugin team.
+ * @license   GPLv3 https://www.gnu.org/licenses/gpl-3.0.html
+ * @link      https://github.com/pluginsGLPI/order
+ * -------------------------------------------------------------------------
+ */
+
+Session::checkLoginUser();
+Session::checkRight('config', UPDATE);
+
+// The 'config' right only carries READ/UPDATE bits, so UPDATE - checked
+// above - is the whole gate; per-item CREATE/PURGE checks would always deny.
+$mapping = new PluginOrderGenerationField();
+
+if (isset($_POST['add'])) {
+    $mapping->add([
+        'itemtype' => (string) ($_POST['itemtype'] ?? ''),
+        'field'    => (string) ($_POST['field'] ?? ''),
+    ]);
+    Html::back();
+}
+
+if (isset($_POST['purge'])) {
+    $mapping->delete(['id' => (int) $_POST['id']], true);
+    Html::back();
+}
+
+Html::redirect(PluginOrderConfig::getFormURL());