+++ setup.php
@@ -30,7 +30,7 @@
 
 use function Safe\define;
 
-define('PLUGIN_ORDER_VERSION', '2.14.1');
+define('PLUGIN_ORDER_VERSION', '2.14.2');
 
 // Minimal GLPI version, inclusive
 define("PLUGIN_ORDER_MIN_GLPI", "11.0.0");